basic regression test for delete route

+++ src/book/book.services.js
@@ -61,6 +61,11 @@ export class BookService {
     }
 
     async deleteBookProfile(id) {
+
+        if (!id) {
+            throw new Error("Book ID must be provided");
+        }
+        
         return await this.BookRepository.delete(id);
     }
 }
